Remove interface and rename impl class
As Atlas not following pure MVC model and interface concepts, I have removed 'MarkerGeneService'  interface and renamed the service impl class as Atlas strictly following the exisitng design patterns.
