+++ src/main/java/uk/ac/ebi/atlas/experimentpage/markergenes/MarkerGeneService.java
@@ -1,16 +1,30 @@
 package uk.ac.ebi.atlas.experimentpage.markergenes;
 
 import com.google.common.collect.ImmutableList;
-
+import org.springframework.stereotype.Service;
 import java.util.List;
+import static com.google.common.collect.ImmutableList.toImmutableList;
 
-public interface MarkerGeneService {
+@Service
+public class MarkerGeneService {
+    private MarkerGenesDao markerGenesDao;
+
+    public MarkerGeneService(MarkerGenesDao markerGenesDao) {
+        this.markerGenesDao = markerGenesDao;
+    }
+
     /**
-     * Get marker genes for an experiment of that organism part.
-     *
-     * @param experimentAccession
+     * Get ImmutableList of marker genes objects for an experiment of that organism part.
      * @param organismPart
-     * @return ImmutableList of CellTypeMarkerGene Objects
+     * @param experimentAccession
+     * @return ImmutableList of CellTypeMarkerGene  Objects
      */
-    ImmutableList<CellTypeMarkerGene> getCellTypeMarkerGeneProfile(String experimentAccession, String organismPart);
+    public ImmutableList<CellTypeMarkerGene> getCellTypeMarkerGeneProfile(String experimentAccession, String organismPart) {
+        List<CellTypeMarkerGene> cellTypeMarkerGenes = markerGenesDao.getCellTypeMarkerGenes(experimentAccession, organismPart);
+        return cellTypeMarkerGenes.stream()
+                .filter(markerGene -> !markerGene.cellGroupValue().equals("Not available"))
+                //'cellTypeWhereMarker' attribute is not feasible to filter 'Not available' cells as it
+                // contains descriptive value ,so selected 'cellType' attribute to filter 'Not available' cells.
+                .collect(toImmutableList());
+    }
 }
